fix(agents): retain prepared music providers through job completion
