+++ src/agents/tools/image-generate-tool.test.ts
@@ -40,7 +40,7 @@ vi.mock("../../config/sessions/session-accessor.js", async (importOriginal) => (
 }));
 
 let imageGenerationRuntime: typeof import("../../image-generation/runtime.js");
-let imageGenerationExecution: typeof import("./image-generate-tool.execution.js");
+let mediaGenerationToolProviders: typeof import("./media-generation-tool-providers.js");
 let mediaGenerationRegistry: typeof import("../../media-generation/registry.js");
 let imageOps: typeof import("../../media/media-services.js");
 let splitMediaFromOutput: typeof import("../../media/parse.js").splitMediaFromOutput;
@@ -354,7 +354,7 @@ describe("createImageGenerateTool", () => {
       };
     });
     imageGenerationRuntime = await import("../../image-generation/runtime.js");
-    imageGenerationExecution = await import("./image-generate-tool.execution.js");
+    mediaGenerationToolProviders = await import("./media-generation-tool-providers.js");
     mediaGenerationRegistry = await import("../../media-generation/registry.js");
     imageOps = await import("../../media/media-services.js");
     ({ splitMediaFromOutput } = await import("../../media/parse.js"));
@@ -368,14 +368,15 @@ describe("createImageGenerateTool", () => {
 
   beforeEach(() => {
     // These fixtures cover routing and limits; the native suite proves actual resource ownership.
-    vi.spyOn(imageGenerationExecution, "acquireImageGenerationToolProviders").mockImplementation(
-      async ({ cfg }) => ({
-        providers: imageGenerationRuntime.listRuntimeImageGenerationProviders({ config: cfg }),
-        assertOpen() {},
-        run: async (run) => await run(),
-        release: async () => {},
-      }),
-    );
+    vi.spyOn(
+      mediaGenerationToolProviders,
+      "acquireImageGenerationToolProviders",
+    ).mockImplementation(async ({ cfg }) => ({
+      providers: imageGenerationRuntime.listRuntimeImageGenerationProviders({ config: cfg }),
+      assertOpen() {},
+      run: async (run) => await run(),
+      release: async () => {},
+    }));
     vi.spyOn(mediaGenerationRegistry, "withImageGenerationProviders").mockImplementation(
       async (cfg, run) =>
         await run(imageGenerationRuntime.listRuntimeImageGenerationProviders({ config: cfg })),
@@ -1280,7 +1281,7 @@ describe("createImageGenerateTool", () => {
 
   it("returns active status for a duplicate image request with the same prompt", async () => {
     const acquireProviders = vi.mocked(
-      imageGenerationExecution.acquireImageGenerationToolProviders,
+      mediaGenerationToolProviders.acquireImageGenerationToolProviders,
     );
     stubImageGenerationProviders();
     vi.stubEnv("OPENAI_API_KEY", "openai-test");
